Optionally search for workflows (#270)
* style: remove trailing whitespace

* docs: update url

The API is now versioned. The link has changed.

* feat: optional search for workflows

When no workflow is set, until now the current workflow has been chosen. But
the workflow we get the artifact from could be another one determined by other
criteria like the branch and the artifact name.

We provide a new option workflow_search, that if set, allows us to search for
the workflow repo-wide according to the other provided data.