Head sha revert (#271)
* fix: get workflows only for commit

In f6b0bac the check on the commit value provided was moved to the body of the
function. Instead directly provide it as parameter of the request as it was
before assuming that this might affect the number of requests sent.

* test: check download without workflow set

This tests searching for a workflow if not provided.