Read workflow_search input as a boolean
In commit f6b0bace624032e30a85a8fd9c1a7f8f611f5737 we added
a check for !workflow_search before using the current workflow,
but with a string this always evaluates to true.

This results in always searching workflows, and if enough
inputs are unspecified it can find the wrong run id.

Signed-off-by: Kyle Harding <kyle@balena.io>